Put modules added. Example provided for rgb.
The put module allows for a string of characters to be provided to a function.
The function should thne set or replace something. It should provide identical
behavior if the same string is passed multiple times.

The example allows for the rgb to be set to any RGB value by simply calling
the function. `put "rgb=255,255,255"` would change the color to white.